update contactpopup.tsx
A contactpopup.tsx page has been created. Add contact is done through that page. When fullName, roleTitle, company, email, phone, department, status are entered and the Add contact button is pressed, the data is sent to the database through the backend. There, the data is sent to the database through companyRepresentative --> contactController.js --> contact.js.