update contactpopup.tsx page
I added styles that match the contactpopup.tsx page. Also, validation was added for the email and phone number.

Add contact button on the all-contacts-view.tsx page is linked to the contactpopup.tsx page.

Add contact button under the quick action in staff-dashboard.tsx is also linked to the contactpopup.tsx page.