develop contacts-table.tsx  page and update contacts-view.tsx
In all-contacts-view.tsx page contactInfo( email and phone) are set direct fields.

changed the id under the Contact type in the contacts-table.tsx page to _id and set it to a string, and changed the role to the position attribute.

Developed the getContactsByEmail function. It developed all the contacts related to the email into the viewed in the contacts-table.tsx page.

developed the code to convert the lists from the database to an array using transformedContacts.

imported the ContactPopup page that needs to be popup when the Add Contact button is click.