update contacts-table.tsx UI and contactpopup.tsx
The contacts-table.tsx page is configured to update the data related to that contact when the Edit Contact button under the DropdownMenuItem is clicked. The contactpopup.tsx page is used for that update. During the update process, an Update Contact button is created instead of the Add contact button on that page.

An endpoint has been added to endpoint.ts for UpdateContact