Update analytics-dashboard.tsx
In the analytics-dashboard.tsx page, the number of My Calls in the database corresponding to the "user_id" is calculated and displayed in that card. Also, the number of contacts added in the 5 days of the week under Weekly Activity was developed and displayed. There, the number of calls is calculated according to the "uploadDate".