update contacts-table.tsx and contactpopup.tsx
The "handleViewDetails" function is set up, which is executed when the "View details" button is clicked on the contacts-table.tsx page. There, the setViewMode opens the ContactPopup() UI (popup).

All the data displayed in the field of the contactpopup.tsx page opened by setViewMode is displayed in read-only format. Also, all the buttons are hidden in view mode.