Update all-contacts-view.tsx  and endpoints.ts
In the all-contacts-view.tsx UI, when you click on the Eye icon under Actions, you will see a dropdown with Completed and Not Complete. When you select the desired one, it will be saved in the database. For this, a function called handleStatusChange was developed.

Also, an endpoint called updateStatus was added to endpoints.ts to connect with the backend.

In the staff-dashboard.tsx UI, Today's Schedule was removed and it was set to display Completed Contacts.