Develope contacts-table.tsx, contactpopup.tsx, note-popup.tsx, endpoints.ts
The NotePopup page is set up to pop up when the "Add Note" button is clicked in the contacts-table.tsx UI.

Added tailwind.css to make the contactpopup.tsx UI look nice.

A new note-popup.tsx UI is created. Here, when the "Add Note" button is clicked in the contacts-table.tsx UI, this note-popup.tsx UI will pop up. Fill in the Notes, Outcome, Next Action, Scheduled Date fields there and finally click the Add Note button to send all that data to the database.

Added the endpoint of the note-popup.tsx page to endpoints.ts.