Create schedule-call-popup.tsx and Update contacts-table.tsx, staff-dashboard.tsx, endpoints.ts
When the Schedule Calls button is clicked in contacts-table.tsx, the schedule-call-popup.tsx UI located in the path "@/components/ui/schedule-call-popup" is configured to popup

A ScheduledCall interface was created in staff-dashboard.tsx and it was made to display the calls that are currently scheduled under Upcoming Scheduled Calls. The fetchScheduledCalls function is used for that. A button was also added to Delete Schedule calls. The Scheduled Calls card is also used to display the current Scheduled Calls.

A schedule-call-popup.tsx UI was added. It allows you to schedule a call to existing contacts. There, add things like Select Contacts, Date, Time, Notes and finally click the Schedule Calls button, all that data is stored in the database.

Added the required endpoints scheduleCalls, getScheduledCalls, deleteScheduledCall to the endpoints.ts page.