Update all-contacts-view.tsx
Added a Calendar to all-contacts-view.tsx to filter contacts. When the desired date is selected from the Calendar, the filteredContacts function is executed. It takes the upload date based on the selected data and returns the contacts for that selected date.

Also added a pagination so that only 10 contacts are displayed in the table.