Update all-contacts-view.tsx and staff-dashboard.tsx
In the all-contacts-view.tsx UI, a button was added to the Actions DropdownMenu of the Table that displays contacts to allow marking as Mark as Hot Lead. A function called handleHotLeadChange was added for that.

Also, a Card was added to the staff-dashboard.tsx UI to display the number of Hot Leads. Also, a card was added to display the calls made to those Hot Leads.